test: accountant e2e contract queries
