+++ wormchain/contracts/tools/package.json
@@ -5,7 +5,7 @@
   "main": "deploy_wormchain.ts",
   "scripts": {
     "deploy-wormchain": "ts-node deploy_wormchain.ts",
-    "test-accountant": "ts-node test_accountant.ts",
+    "test-accountant": "jest test_accountant.ts --verbose",
     "test-wormchain": "ts-node test_wormchain.ts",
     "deploy-and-test": "npm run deploy-wormchain && npm run test-wormchain"
   },
@@ -26,6 +26,9 @@
   },
   "devDependencies": {
     "@types/elliptic": "6.4.14",
+    "@types/jest": "29.4.0",
+    "jest": "29.4.1",
+    "ts-jest": "29.0.5",
     "ts-node": "10.9.1",
     "typescript": "4.9.4"
   }
