+++ wormchain/contracts/tools/tsconfig.json
@@ -1,5 +1,6 @@
 {
-    "compilerOptions": {
-        "resolveJsonModule": true
-    }
+  "compilerOptions": {
+    "resolveJsonModule": true,
+    "esModuleInterop": true
+  }
 }
